+++ src/applications/content-upload-control-app/upload-list/upload-list-table.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input, Output, EventEmitter, ViewChild, ChangeDetectorRef, AfterViewInit } from '@angular/core';
 import { AppLocalization } from '@kaltura-ng/kaltura-common';
-import { NewUploadFile } from '../../kmc-upload-app/kmc-upload-app.service';
+import { KmcUploadManagementService, UploadFileData } from 'app-shared/kmc-shell';
 
 @Component({
   selector: 'kUploadListTable',
@@ -9,7 +9,7 @@ import { NewUploadFile } from '../../kmc-upload-app/kmc-upload-app.service';
 })
 export class UploadListTableComponent implements AfterViewInit {
   @Input()
-  set uploads(data: Array<NewUploadFile>) {
+  set uploads(data: Array<UploadFileData>) {
     if (!this._deferredLoading) {
       // the table uses 'rowTrackBy' to track changes by id. To be able to reflect changes of entries
       // (ie when returning from entry page) - we should force detect changes on an empty list
@@ -23,7 +23,7 @@ export class UploadListTableComponent implements AfterViewInit {
     }
   }
 
-  @Input() selectedUploads: Array<NewUploadFile> = [];
+  @Input() selectedUploads: Array<UploadFileData> = [];
 
   @Output()
   sortChanged = new EventEmitter<any>();
@@ -32,10 +32,10 @@ export class UploadListTableComponent implements AfterViewInit {
   onSelectedEntriesChange = new EventEmitter<any>();
 
   @Output()
-  onCancelUpload = new EventEmitter<NewUploadFile>();
+  onCancelUpload = new EventEmitter<UploadFileData>();
 
   private _deferredUploads: any[];
-  public _uploads: Array<NewUploadFile> = [];
+  public _uploads: Array<UploadFileData> = [];
   public _deferredLoading = true;
   public _emptyMessage = '';
 
@@ -61,14 +61,15 @@ export class UploadListTableComponent implements AfterViewInit {
     }
   }
 
-  public _relatedTableRowStyle(rowData: NewUploadFile): string {
-    if (rowData.uploadFailure) {
-      return 'has-error';
-    }
-
-    if (rowData.removing) {
-      return 'removing';
-    }
+  public _relatedTableRowStyle(rowData: UploadFileData): string {
+    // TODO [kmcng] original data provided from service doesn't have those properties
+    // if (rowData.uploadFailure) {
+    //   return 'has-error';
+    // }
+    //
+    // if (rowData.removing) {
+    //   return 'removing';
+    // }
 
     return '';
   }
