+++ src/shared/kmc-shell/index.ts
@@ -2,3 +2,4 @@ export * from './providers/index';
 export * from './components/index';
 export * from './kmc-shell.module';
 export * from './auth';
+export * from './kmc-upload-management.service';
