+++ src/shared/kmc-shell/kmc-shell.module.ts
@@ -9,6 +9,7 @@ import { BrowserService } from "./providers/browser.service";
 import { AppContainerComponent } from "./components/app-container/app-container.component";
 import { ReleaseNotesComponent } from "./components/release-notes/release-notes.component";
 import { EntryTypePipe } from 'app-shared/kmc-shell/pipes/entry-type.pipe';
+import { KmcUploadManagementService } from 'app-shared/kmc-shell/kmc-upload-management.service';
 
 @NgModule({
     imports: <any[]>[
@@ -28,6 +29,7 @@ import { EntryTypePipe } from 'app-shared/kmc-shell/pipes/entry-type.pipe';
         EntryTypePipe
     ],
     providers: <any[]>[
+
     ]
 })
 export class KMCShellModule {
@@ -43,7 +45,8 @@ export class KMCShellModule {
             ngModule: KMCShellModule,
             providers: <any[]>[
                 BrowserService,
-                AppShellService
+                AppShellService,
+                KmcUploadManagementService
             ]
         };
     }
